[fake_ddk] Clarify which parts of Bind are public vs protected.
In particular, we leave members that are expected to be used by tests
as public, while making those that are expected to be overridden by
subclasses (but not directly called into) as protected.

This requires us to list the entire set of global DDK functions as
"friend" which isn't ideal, but perhaps makes it clearer to readers
the expected use of the Bind class.

Change-Id: Ibd9decbd77cd6e3b1512796cf0451a37c9d9d8bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378956
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>