[amlogic_decoder][h264] Use pull model for input
Storing a list of NAL units in the decoder is good for simple tests, but
in production it could lead to a large queue of undecoded data. To fix
this, start moving to a pull model, where the decoder can request more
data using a FrameDataProvider.

Bug: 13483

Change-Id: Ia5a3cb3cab3af97a6cced4c1b6976be6fe9d3452
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379430
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>