[fshost] Increase pkgfs connect deadline to 20s
pkgfs has had many changes recently, which increase the expected time to
connect to fshost. We are increasing the timeout to 20s to accomodate
for these changes, and reduce the likelihood that fshost times out and
the system fails to boot.

Bug: 49317
Test: fx qemu
Change-Id: I7b04cf7f84266d63b5194c6c351ea6dc0a61edea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377355
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>