[ledger] Work around flake in LedgerEndToEndTest.ClearPage
There seems to be a bug at the storage layer where readdir() will return
nullptr and leave errno to 0 (indicating end-of-stream) immediately
after fdopenat(), without even returning a dirent for the current (".")
directory. It is suspected that this happens if the directory is deleted
concurrently, but this part of the investigation is left to a follow-up
bug.

This commit avoids a crash in that case, instead behaving as if the
directory wasn't found at all (ie. as if errno had been set by
readdir()). In principle this would still lead to flakes in the test,
but in practice this makes them disappear because the issue happens
during a recursive descent, and it is unlikely -- or even maybe
impossible given the specifics of the scenario -- to be triggered on the
critical path leading to the looked-for directory; it is believed that
the disappearing directory is always a sibling of the interesting one.

This commit also re-enables the test now that the flake is eliminated,
and takes the opportunity to port change
I4bed327b2d19e156c8daebfe7e563387da191191 to ledger's copy of
ReadDirContentsAt.

Bug: 49344
Test: fx test -o ledger_e2e_local -- --gtest_filter=LedgerEndToEndTest.ClearPage --gtest_repeat=-1 --gtest_break_on_failure
Change-Id: I78bc6e810c13da3c04b229b5da2d2e833b2eeb5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379674
Reviewed-by: Benjamin Lerman <qsr@google.com>
Testability-Review: Benjamin Lerman <qsr@google.com>
Commit-Queue: Gabriel Kerneis <kerneis@google.com>