[policy] Remove cmx internal features.
This changeset removes cmx internal features, unifying around features
as a single source of truth for manifests. Prior to this change
global-data would be added to archivist.cmx and observer.cmx at build
time. This made this code path more difficult to audit.

With the introduction of the appmgr policy checker we have runtime
checks that enforce this behavior and provide a true security boundary
anyway so this feature is no longer required.

This CL updates archivist.cmx to include global-data and updates the
integration tests to validate that the other usages are blocked.

Bug: 49830
Change-Id: I488290ee41c4843e33e8e42e83b05e48c6c4a353
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379509
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Testability-Review: Adam Perry <adamperry@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>