[modular] Instrument alignment of session creation requests.
This is a precursor to continuing the removal of account manager from
modular. We compare all requests to create a session with the existing
session directories to assess whether we could have determined the
account id using these directories. The results are output to colbalt
and logs.

This also tidies up a few compiler and FIDL errors on the unittest for
session context.

Bug: 45946,48782
Change-Id: I0bf96628845d6e4eade804f8c44e4a5f2237a981
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376314
Reviewed-by: Gabe Schine <thatguy@google.com>
Testability-Review: Gabe Schine <thatguy@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>