[gidl][go] Generate expressions instead of builder statements
This CL refactors the Go backend of GIDL to produce a single expression
for each value instead of emitting statements to build the value. This
results in a simpler implementation and generates smaller output files.

For example, to build the vector<string> value ["one", "two", "three"],
it used to produce this:

    v1 := "one"
    v2 := "two"
    v3 := "three"
    v4 := []string{v1, v2, v3}

Now, it produces a single expression:

    []string{"one", "two", "three"}

The Rust and Dart bindings already use this approach. With this CL, only
the C++ backends use the statement approach.

Test: fx test fidl_go_conformance -- -test.v
Change-Id: I47301c9eaa3b7134adeeee9abf5b09d87b43a2b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379598
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Benjamin Prosnitz <bprosnitz@google.com>