[rust][build] Deprecate rust_panic
In preparation for shipping a prebuilt libstd for fuchsia with
panic=abort.

Alternatives to removing rust_panic are shipping a separate prebuilt
libstd for panic=abort and panic=unwind, and building libstd in-tree.
This seems like the simplest option by far. Building libstd in-tree
might be nice, but it's unclear (to me) if we want that.

Bug: 49860
Change-Id: If348edca92b50b37c04c634d1b74f84667dcdeda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379524
Reviewed-by: Didrik Nordstrom <dnordstrom@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>