[sl4f][benchmarks] Fix filename check: avoid assert(), which is disabled
Convert an assert() to an explicit throw so that the check for a
filename extension isn't skipped.

assert() is not always enabled in Dart code, and apparently our Dart
code runs with assert() disabled.  (This makes assert() mostly
useless, similar to how it should be avoided in C/C++.  Most of the
time we want a CHECK() rather than a DCHECK().)

Take the opportunity to introduce a _removeSuffix() function here.

Bug: 45804
Test: added unit test; run locally with:
  fx set terminal.x64 --with sdk/testing/sl4f/client:tests
  fx run-host-tests sl4f_client_tests
Change-Id: I5e1c7ac327704728bf1f974d0b16f88416d6448a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379501
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>