[flatland] Add pixel_scale to LayoutInfo table.
Pixel scale for a particular link is computed as the product of the
scale portion of the link handle's global matrix and the
display-specified pixel scale.

Because link size to logical size ratios are included in the global
matrices, extracting the scale from the global matrix includes all
relevant scales that lead to the link handle.

The display-specified pixel scale is currently a separate value, but
could be included in the global matrices in the future depending on
how a display connects to a Flatland hierarchy. For example, if the
display owns a single-Transform Flatland instance that links to the
actual root Flatland instance, applying a scale matrix to that
Transform would eliminate the need for a specific argument for the
display.

TEST=fx test fuchsia-pkg://fuchsia.com/scenic_tests#meta/flatland_unittests.cmx

MULTIPLY: `[
  {
    "name": "scenic_tests",
    "os": "fuchsia",
    "total_runs": 40
  }
]`

Change-Id: I44e54cc78a9bb808c2dcf00f6507368ca7542d3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378563
Commit-Queue: Evan Birenbaum <birenbaum@google.com>
Testability-Review: Evan Birenbaum <birenbaum@google.com>
Reviewed-by: Adam Gousetis <adamgousetis@google.com>
API-Review: Adam Gousetis <adamgousetis@google.com>