[gn sdk] Add GN SDK emulator tests for OSX on Linux and fix fxr/376644
The emulator for OSX was updated in fxr/376644, which added a new
upscript. However, this needs to be copied over by the GN SDK, and
there was nothing that caught this, since it only runs on OSX.
A new set of tests have been added that sets up uname output that
looks like OSX so that we can run these tests on Linux machines,
which would have caught the change in fxr/376644. The original
zircon qemu-ifup-macos script was forked into a new version just
for fx emu to remove any dependencies.

Change-Id: I9deefce477094b17606d9e47bb9cf30c126b0c53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378089
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Wayne Piekarski <waynepie@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>