[microbenchmarks] Do multiple process runs in SL4F test for parity with non-SL4F
This is in preparation for switching the perfcompare trybot to using
the SL4F-based wrapper
(src/tests/end_to_end/perf/test/microbenchmarks_test.dart) instead of
the older non-SL4F-based wrapper (src/tests/benchmarks/benchmarks.cc).

Change microbenchmarks_test.dart to do multiple process runs of
fuchsia_microbenchmarks similar to how benchmarks.cc does that.  The
reasons are described in the comments, which are based on the comments
in benchmarks.cc.

A difference from benchmarks.cc is that we do the extra process runs
unconditionally, even when the test is not running in the context of
perfcompare.  This is for simplicity, and the extra running time is
not much.  (There is no flag plumbed through to indicate whether the
test is running on a perfcompare builder.  We could check the
BUILDER_NAME env var, but we don't really want to conditionalise on
infra builder names.)

Bug: 49280
Test: on CQ: fuchsia-x64-perfcompare-dawson_canyon-sl4f
 * Confidence intervals *should* get narrower, but this is currently
   confounded somewhat by issue 49784.
Change-Id: I1b59fca7e10222df51924180388525ac5f0d6ff1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380082
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>