[sshutil] Have ssh commands be slices of strings
sshutil previously represented commands to run over ssh as strings with
whitespace-separated arguments. This is the form in which the Go ssh
library takes commands, but it's inconsistent with Go stdlib functions
like exec.Command(), as well as the SSHRunner library that botanist and
testrunner use.

I also reformatted many Run() calls of the form:

  err := client.Run("command")
  if err != nil { ... }

to be of the form:

  cmd := []string{"command"}
  if err := client.Run(cmd); err != nil { ... }

This is consistent with the Go style guide. I kept `cmd` declarations on
a separate line to keep the `if` line relatively short.

Bug: 48042
Change-Id: Ie124f257e03a81ffdf2d06c035ea495159c2279b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380073
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>