[scenic] Enable Vulkan pipeline cache warming in Scenic.
Step 9 in a series of CLs to build Vulkan pipelines at startup, rather
than lazily as needed.  This CL enables the functionality in Scenic,
woo-hoo!

The goal is to avoid any jank due to lazy pipeline creation in the
middle of a frame.  However, if lazy pipeline creation is attempted
accidentally it will succeed, and a Cobalt event is logged.

Fixed: 44084

Change-Id: I791b3bf424729a93c79d60b7899358c9e0758b5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/358888
Testability-Review: Yilong Li <liyl@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Yilong Li <liyl@google.com>