[gn sdk]Updating test driver for relative paths
This change is in preparation for runnning the bash script tests
as part of the "normal" CQ. It makes locating the paths of the
scripts to run to be either absolute paths or relative paths where
they are relative to the executable.

Testability: These are tests being updated, so the tests should still
pass, and no new tests that use the relative functionality have been
added yet.

Change-Id: Ie87e03234623a0eb6b3a9a41f459256f1c9a4f40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380181
Reviewed-by: Jonathan Koren <jdkoren@google.com>
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>