[ffx][daemon] Add wait for RCS connection.
This adds an implementation to the `Target` class that will allow a
caller to wait for RCS to start given a number of attempts and delays.

This is not an ideal solution, as it would make more sense to use
something like std::sync::CondVar, but async_std and futures don't
support these kinds of synchronization structures.

Fixed: 49659
Change-Id: I82366b64dd1a9b71fdf9e218744fe76c8ef63a50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379072
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Testability-Review: James Tucker <raggi@google.com>