[gn sdk]Updating tests to run out of src tree
This moves the sourcing of the library scripts and referencing
test data to be relative to the scripts under test vs. the source
tree.

This allows the tests to be run as part of the "normal" CQ.

Testability: This is a refactor of tests, no functional changes
so everything should work as now with no regressions.

Change-Id: Idb65fa27300722c0f2e16a2b62fa63ef216baf71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380357
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>