[trace_processing] Add frame latency to flutter_frame_stats
Adds a metric that measures the time between flutter starting to produce
a frame and it getting displayed on the device. This is intended to
provide an alternative to input latency metrics for use cases where the
output is driven by a programmatic state change rather than a user
input.

Bug: 49925

Change-Id: I639b99d91ff6b3102af10d1d79aee21fbcb89a1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375002
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>