[brcmfmac] Fix for test flakiness.
Exit the timer handler as soon as the number of iterations are
done. Also stop the timer as soon as the async indication is
received. Also stop the timer in the handler once the required
number of iterations are done.

MULTIPLY: `[
  {
    "name": "timer_test",
    "total_runs": 50
  }
]`

Change-Id: I4b22c2cc51959d5d0d1d0d6f6312621b1d95a28d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380279
Commit-Queue: Karthik Krishnaswami <karthikrish@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>