[chunked-compress] Introduce chunked compressor library
This change introduces "chunked-compress", a library for creating
compressed archives that support decompressing sub-segments of. This is
intended to be used to support demand paging of compressed files, for
example in Blobfs, without having to decompress the entire file.

The format is fairly simple: A header, a seek table, and a number of
frames. Each frame is a regular ZSTD archive.

Detailed documentation can be found in "archive-header.h".

This change only adds the compressor (decompressor is in a child CL).

Test: Extensive unit/manual testing.
Change-Id: I1d929fdd08aaca8cefe10550b2f3aede5234fd65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378774
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>