[bt][audio-output] Drop audio socket with stream
When the audio stream was dropped, the SoftPcmOutput thread
handling the audio device would continue on, with the audio system
unawares that no one was listening.  Add a handle so when the audio
output stream is dropped, the background task is also stopped.

This drops the audio device channel and any RingBuffer channel
established as well.

Fixed: 43812
Test: added soft_pcm_audio_should_end_when_stream_dropped
      fx test fuchsia-audio-device-output-tests
      manually tested with bt-a2dp-source output
Change-Id: I34307f6cc835a336c14a8a2d32aea1c853e76e67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380436
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>