[swd-e2e-tests] Consolidate test paving logic
This extends `device.Client to take advantage of `Build` by adding a
`device.Client.Pave` method that paves the device to that specific
build.

Furthermore, this extracts the test paving logic into
`system-tests/pave`. This module supports both rebooting to recovery, as
well as OTA-ing to recovery, so this patch also adds support for this to
the reboot and tracking tests.

Test: Verified OTA tests still pass with this CL applied.

Change-Id: I8b4d997f98621d7ae63f20b33af0e542491b2dd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378869
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Ben Keller <galbanum@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>