[qemu] Remove QEMU exit code check
It used to be the case that the infrastructure needed to configure QEMU
on x64 to exit with status code 31 in some cases - and for it to not regard
such a nonzero exit as failure. This requirement is no longer true, so
it's time to delete some code.

Change-Id: If6d50666f0006e32c95a1a345f6c5e222f6701ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380574
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>