[bt][hci] Refactor driver to fix an edge case.
Refactor this driver both to fix an edge case error and to also
cleanup some of the code in general.

The edge case error goes like this.  This driver is an implementation
of the HCI parsing layer for bluetooth sitting on top of a more
generic serial UART interface.  The UART can currently handle a single
packet write operation in flight at once.  The HCI driver layer seems
to be buffering its UART write jobs using its client's channels
instead of draining the client write requests from the channels in to
driver local memory.  It cannot read from the channel in order to
write to the UART unless the UART is able to take another write job.

Because of this, the HCI driver layer has an interlock mechanism which
is intended to keep only one write operation in flight to the UART
driver at once.  Any error during an attempt to write to the UART
driver was considered to be fatal and the driver will attempt to shut
down.

The edge case shows up in the HCI driver's main work thread loop.  It
was always waiting for one or more client's channel to have work to
do, and then (after that wait) it was waiting again on the event it
had set up to indicate that the UART was write-able.  After that, it
would go ahead and service the first write request (if any) from each
of its pending client channels.

The problem is that if _multiple_ client channels have pending write
operations at the same time, it would write the first operation to the
UART, and then attempt to write the second operation immediately
afterwards.  This would fail, and cause the driver to panic and shut
down.

This is not the only potentially problematic thing going on with the
driver right now.  Here is a list of some of the refactoring changes
meant to address some of these issues.

1) Restructure to main work loop to be a bit mode direct about how it
   figures out what to do.  Always make a list of the events it cares
   about, then wait on them at the start of each loop.
2) Consolidate the multiple events which were being kept around to
   wake up the driver in case something external had changed.  Replace
   these all with a single wakeup event which anyone can poke.  Any
   time the thread is poked, it will immediately re-evaluate its
   priorities before proceeding.
3) Keep track of whether or not we think that our UART is writeable or
   not based on an internal state variable and do not allow ourselves
   to attempt to write to the UART if we don't think that we can right
   now.
4) Police the way that client channels get closed.  There are at least
   3 different thread contexts in play here right now.  The DDK
   framework, the HCI work thread, and the UART callback thread.  The
   HCI work thread, once started, needs to be the owner of the
   lifecycle of these handles as it needs to wait on the handles
   outside of any lock.  If anyone else attempts to close the channels
   (like the UART callback thread), it runs the risk of letting the
   work thread wait on an invalid handle which would be a fatal job
   policy violation for a driver.  So, once the driver is up and
   running, make sure that it is the only one who ever closes channels
   in response to PEER_CLOSED or a write error on the channel.  During
   shutdown, make sure that the thread has exited completely before
   closing the channels.
5) Fix at least one place where the error handing looks like it was
   damaged by some copy-pasta.  In the handler called by the UART when
   there was newly read data, a failure to write to the CMD channel
   would goto a bailout tag which would end up closing the ACL channel
   instead of the CMD channel (although, closing this channel from the
   UART callback thread was never safe to begin with).
6) If we have a UART read error, consider it to be fatal and
   immediately start to shut down, instead of simply failing to queue a
   new read job.  While these errors should never happen, if they do,
   we were previously putting our driver into a zombie state where it
   would never read again instead of just shutting down and thereby
   notifying our users that the show was over.

Moving forward, I think it would be a good idea to consider making
some deeper changes in how this code is structured, both to increase
performance and to decrease the potential for mistakes.  My
suggestions are as follows.

1) Convert this driver from C to C++.  There are many places in this
   code which would benefit greatly from proper application of object
   oriented patterns and the reduction in code duplication that could
   be attained by doing so (the client channels are prime examples of
   things which should be their own objects).  Making use of static
   thread analysis would be highly beneficial as well, but is
   difficult to do in just C.  Finally, there are many many places
   where RAII patterns and helpers could be used effectively.  In
   particular, AutoCalls instead of gotos, managed pointers instead of
   remembering to free in error paths, and zx::handle management
   instead of manual management.
2) Eliminate the use of wait_many and prefer the use of a port
   instead.  Wait many can be inefficient, ports offer ways to more
   easily post messages to a thread, and to associate information with
   a handle being waited on instead of needing to look up and match a
   handle to a job context upon wakeup.
3) Change the UART driver to allow more than one write in-flight at
   once, and also expose explicit visibility of the write-ability of
   the UART to the HCI level.  Interlocking an needing to wake two
   threads every time a write operation is finished is probably not
   ideal for performance.
4) Keep the client channels drained.  Not processing messages pending
   in a client channel is just not a good idea as it is relying on the
   kernel to act as your buffer.  Instead, the channels should be read
   ASAP into a queue of pending work in the driver.  If the amount of
   work from a single client is becoming unreasonable, the driver
   should simply close the connection to the client and purge the work
   queue instead of letting the channel suck up all of the kernel
   memory.

Test: Stress testing using an on-going A2DP link to a pair of
      headphones.  Previously, this test case used to fail in an hour
      or two when the edge case mentioned above was hit.
Change-Id: Idf844b6bc9b8931919b88d774d6ca0a2faae6096
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379770
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>
Testability-Review: Jeff Belgum <belgum@google.com>