[build] forbid v2 components in cache and universe
This change prevents CFv2 components from being in cache and universe
packages. We are doing this because if any v2 components are in cache or
universe, they _won't_ be able to launch. This is because v2 components
are launched using /pkgfs/packages, and we recently changed
/pkgfs/packages so that it only shows base packages. Context:
fxrev.dev/374938.

Fixed: 48948
Change-Id: I6b25bacd28a01c5653daa5371e2d675651aefd09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379153
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Jake Ehrlich <jakehehrlich@google.com>