[storage][volume-image]: Update BUILD.
1. Make test run in host side as well.
2. Include the tests as part of storage tests.
3. Move from executable to test(required for host tests).
4. Added path utils that obtains the base path for looking up schemas.
5. Fix outstanding OOB issue that asan detected.

w.r.t. (4) -> On fuchsia whe have a sandboxed state, on the other hand
on host this is not the case. We know that the dependencies will be
relative to the executable path, but the CWD is undefined. The
recommended approach is figure out the executables path, and calculate
path form there.

This is how zxdb does it. Could depend on their host util, but would
pull a lot of dependencies with that. So forked their impl.

TEST=storage-volume-image-tests

Change-Id: I574fbae09b59e878bb24ae05c81a3695cd06ea50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380520
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Richard Chen <tianruichen@google.com>