[net-svc]Add ipv6 ULA and SLAAC to zedboot
Assign a Unique Local Address to zedboot to use for paving/netboot.
There are issues, around chromebooks in particular, with Link Local
addresses.

The unique-ness here is taken from the MAC address. Conflicts could c
ome if we are using devices with the same mac but that seems unlikely.
This address is designed to not be routed so that shouldn't be an issue either.

This also adds a router advertisement message with our netboot
advertisement. The RA packet will cause connected devices to add an ipv6
address on the connected link that can talk to our ULA.

Information about SLAAC and the router advertisement packet can be found
here: http://go/rfc/4861

Information about ULA addresses here: http://go/rfc/4193

Tested: Tested on a sherlock and astro connected to a personal linux laptop and
then a chromebook (lenovo c630). Works as expected on both and both
configure to talk to the ULA.

Change-Id: I99d4d1471164fc89371c64d0b73be0ba200056f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374678
Commit-Queue: Ed Coyne <edcoyne@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>