[logs] Type-safe severity parsing.
In order to support a new severity scale, we need to support two
sets of ordinals in the wire format simultaneously. All of the
meaningful *readers* of this value are in-tree, so we plan to do
the migration by rewriting old->new in the Archivist.

To be able to rewrite the current/old severity values into the new
severity values, we need to parse the current values (which we
previously didn't do at all).

Bug: 49474
Change-Id: Ifa54eb34e838a3b0c4f08756f864fc7ae13f2a3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378973
Commit-Queue: Adam Perry <adamperry@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>