[swd-e2e-tests] Use ctx in sl4f; wrap major phases in timeouts
Previously, the e2e tests used contexts to provide a timeout mechanism,
but unfortunately this relies on contexts being used in every
potentially long running call. We missed one in sl4f when we make an
http request. This hung in a test, and resulted in the whole test
hanging.

This addresses this hang in two ways. First, we pass the context to the
sl4f http client, so if the context gets canceled, the http request
should also get canceled.

Second, we extract out the pattern of "run a cancelable function in a
goroutine, and wait for it to finish, or for the context to cancel. We
then use this functionality to time out the major phases of each test in
case they hang. Finally, we migrate all the manual implementations of
this pattern to use this function to reduce code duplication.

Test: Ran tests, verified they still pass locally
Change-Id: Ib6e900b104e53242631dbb26b8912549919103ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380762
Reviewed-by: Alex Legg <alexlegg@google.com>
Testability-Review: Alex Legg <alexlegg@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>