[debugger] Try to fix racy test.
Try to fix a flake in the debugger-test. The problem was that waiting
for the SUSPENDED notification would occationally result in a thread
that was RUNNING. The best theory is that the thread notifications are
racing from the test's previous unsuspend of the thread.

This change forces the target thread to be in a RUNNING state before
waiting for it to be SUSPENDED again.

Bug: 48822

Change-Id: I198b735c5cc6ee268b2c46ff519d7806e92f3c4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380166
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>