[recovery] Make recovery UI a Carnelian app
Make the recovery UI as full Carnelian app,
instead of a mix of direct use of the frame
buffer and Canvas. This is needed for two
reasons; first is that canvas will be going away
in favor of the render interface, second is that
the recovery UI needs to be rotated on some
hardware and the render interface makes that much
easier.

This port should not be a performance regression
from the canvas-based code, but it could be
greatly improved by not redrawing the entire
screen on each vsync.

Testing: ran on NUC and tested hitting the HTTP
server with curl, ran unit tests.

Change-Id: I5b12b8cd4410f171f256c3a77512ab60892e1d48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378565
Commit-Queue: Rob Tsuk <robtsuk@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Testability-Review: David Reveman <reveman@google.com>